+++ Darkages.Server/Storage/locales/debuffs/debuff_morcradh.cs
@@ -30,7 +30,7 @@ namespace Darkages.Storage.locales.debuffs
         public override void OnApplied(Sprite Affected, Debuff debuff)
         {
             if (AcModifer.Option == StatusOperator.Operator.Add)
-                Affected.BonusAc -= (sbyte)AcModifer.Value;
+                Affected.BonusAc -= AcModifer.Value;
 
             base.OnApplied(Affected, debuff);
         }
@@ -38,7 +38,7 @@ namespace Darkages.Storage.locales.debuffs
         public override void OnEnded(Sprite Affected, Debuff debuff)
         {
             if (AcModifer.Option == StatusOperator.Operator.Add)
-                Affected.BonusAc += (sbyte)AcModifer.Value;
+                Affected.BonusAc += AcModifer.Value;
 
             base.OnEnded(Affected, debuff);
         }
