+++ Darkages.Server/Types/Item.cs
@@ -235,9 +235,9 @@ namespace Darkages.Types
             {
                 //Inverted Modifier
                 if (Template.AcModifer.Option == StatusOperator.Operator.Add)
-                    client.Aisling.BonusAc += (sbyte)Template.AcModifer.Value;
+                    client.Aisling.BonusAc += Template.AcModifer.Value;
                 if (Template.AcModifer.Option == StatusOperator.Operator.Remove)
-                    client.Aisling.BonusAc -= (sbyte)Template.AcModifer.Value;
+                    client.Aisling.BonusAc -= Template.AcModifer.Value;
 
                 client.SendMessage(0x03, string.Format("E: {0}, AC: {1}", Template.Name, client.Aisling.Ac));
                 client.SendStats(StatusFlags.StructD);
@@ -418,9 +418,9 @@ namespace Darkages.Types
             {
                 //Inverted
                 if (Template.AcModifer.Option == StatusOperator.Operator.Add)
-                    client.Aisling.BonusAc -= (sbyte)Template.AcModifer.Value;
+                    client.Aisling.BonusAc -= Template.AcModifer.Value;
                 if (Template.AcModifer.Option == StatusOperator.Operator.Remove)
-                    client.Aisling.BonusAc += (sbyte)Template.AcModifer.Value;
+                    client.Aisling.BonusAc += Template.AcModifer.Value;
 
                 client.SendMessage(0x03, string.Format("E: {0}, AC: {1}", Template.Name, client.Aisling.Ac));
                 client.SendStats(StatusFlags.StructD);
